+++ include/field_compute/computation_from_fields.hpp
@@ -3,118 +3,34 @@
 
 #include <type_traits>
 #include "../field/field_accessor.hpp"
-#include "../field_list/field_list.hpp"
 #include "../lib/metaprog/fixed_string_list.hpp"
 
 
 using namespace s2s_literals;
 
 namespace s2s {
-template <auto callable, typename return_type, typename struct_field_list_t, field_name_list field_list>
-struct is_invocable;
-
-template <auto callable, 
-          typename return_type, 
-          typename struct_field_list_t, 
-          fixed_string... req_fields>
-struct is_invocable<callable, 
-                    return_type, 
-                    struct_field_list_t, 
-                    fixed_string_list<req_fields...>> {
-  static constexpr bool res = 
-      std::is_invocable_r_v<
-        return_type, 
-        decltype(callable),
-        decltype(struct_field_list_t{}[field_accessor<req_fields>{}])...
-      >;
-};
-
-template <auto Func, 
-          typename R, 
-          typename struct_fields, 
-          typename req_fields>
-concept can_eval_R_from_fields = 
-  is_invocable<
-    Func, 
-    R, 
-    struct_fields, 
-    req_fields
-  >::res;
-
-
 template <auto callable, typename R, field_name_list Fs>
 struct compute;
 
 // todo: static_vector over fixed_string list?
 template <auto callable, typename R, fixed_string... req_fields>
 struct compute<callable, R, fixed_string_list<req_fields...>>{
-  template <typename... fields>
-    requires (can_eval_R_from_fields<
-                callable, 
-                R,
-                struct_field_list_impl<fields...>,
-                fixed_string_list<req_fields...>>)
-  constexpr auto operator()(const struct_field_list_impl<fields...>& flist) const -> R {
-    return callable(flist[field_accessor<req_fields>{}]...);
-  }
+  static constexpr auto C = callable;
+  using return_type = R;
+  using in_arg_types = fixed_string_list<req_fields...>;
 };
 
 template <auto callable, field_name_list req_fields>
 using eval_bool_from_fields = compute<callable, bool, req_fields>;
+
 template <auto callable, field_name_list req_fields>
 using predicate = compute<callable, bool, req_fields>;
+
 template <auto callable, field_name_list req_fields>
 using eval_size_from_fields = compute<callable, std::size_t, req_fields>;
+
 template <auto callable, field_name_list req_fields>
 using parse_if = eval_bool_from_fields<callable, req_fields>;
-
-template <typename T>
-struct is_compute_like;
-
-template <auto callable, typename R, fixed_string... req_fields>
-struct is_compute_like<compute<callable, R, with_fields<req_fields...>>> {
-  static constexpr bool res = true;
-};
-
-template <typename T>
-struct is_compute_like {
-  static constexpr bool res = false;
-};
-
-template <typename T>
-inline constexpr bool is_compute_like_v = is_compute_like<T>::res;
-
-template <typename T>
-struct is_eval_bool_from_fields;
-
-template <auto callable, fixed_string... req_fields>
-struct is_eval_bool_from_fields<compute<callable, bool, with_fields<req_fields...>>> {
-  static constexpr bool res = true;
-};
-
-template <typename T>
-struct is_eval_bool_from_fields {
-  static constexpr bool res = false;
-};
-
-template <typename T>
-inline constexpr bool is_eval_bool_from_fields_v = is_eval_bool_from_fields<T>::res;
-
-template <typename T>
-struct is_eval_size_from_fields;
-
-template <auto callable, fixed_string... req_fields>
-struct is_eval_size_from_fields<compute<callable, std::size_t, with_fields<req_fields...>>> {
-  static constexpr bool res = true;
-};
-
-template <typename T>
-struct is_eval_size_from_fields {
-  static constexpr bool res = false;
-};
-
-template <typename T>
-inline constexpr bool is_eval_size_from_fields_v = is_eval_size_from_fields<T>::res;
 } /* namespace s2s */
 
 #endif // _COMPUTATION_FROM_FIELDS_HPP_
