+++ include/field_size/field_size_deduce.hpp
@@ -6,20 +6,12 @@
 #include "../field_list/field_list.hpp"
 #include "../field_size/field_size.hpp"
 #include "../field_compute/computation_from_fields.hpp"
+#include "../field_compute/computation_from_fields_impl.hpp"
 
 
 namespace s2s {
 template <typename T>
 struct deduce_field_size;
-
-template <std::size_t N>
-struct deduce_field_size<field_size<fixed<N>>> {
-  using field_size_type = field_size<fixed<N>>;
-
-  constexpr auto operator()() -> std::size_t {
-    return field_size_type::size_type_t::count;
-  }
-};
 
 template <fixed_string id>
 struct deduce_field_size<field_size<field_accessor<id>>> {
@@ -36,7 +28,7 @@ struct deduce_field_size<field_size<size_from_fields<callable, req_fields>>> {
   using field_size_type = compute<callable, std::size_t, req_fields>;
   template <typename... fields>
   constexpr auto operator()(const struct_field_list_impl<fields...>& struct_fields) -> std::size_t {
-    return field_size_type{}(struct_fields);
+    return compute_impl<field_size_type>{}(struct_fields);
   }
 };
 } /* namespace s2s */
