+++ include/type_deduction/type_deduction_clause.hpp
@@ -2,7 +2,7 @@
 #define _TYPE_DEDUCTION_CLAUSE_HPP_
 
 #include "type_deduction_tags.hpp"
-#include "../field_compute/computation_from_fields.hpp"
+#include "../field_compute/computation_from_fields_traits.hpp"
 
 
 namespace s2s {
@@ -12,7 +12,7 @@ concept evaluates_to_bool = is_eval_bool_from_fields_v<T>;
 
 template <evaluates_to_bool eval, type_tag_like T>
 struct branch {
-  static constexpr auto e = eval{};
+  using expression = eval;
   using type_tag = T;
 };
 } /* namespace s2s */
