+++ include/type_deduction/type_deduction_switch_impl.hpp
@@ -13,9 +13,8 @@ namespace s2s {
 
 // todo check if case and eval result match in terms of types
 template <std::size_t idx, typename match_case>
-struct type_switch_impl {
-  constexpr auto operator()(const auto& v) const -> 
-    type_deduction_idx
+struct evaluate_switch_impl {
+  constexpr auto operator()(const auto& v) const -> type_deduction_idx
   {
     if(v == match_case::value) return idx;
     else return std::nullopt;
@@ -23,32 +22,29 @@ struct type_switch_impl {
 };
 
 template <typename... match_cases>
-struct type_switch_helper {
+struct evaluate_switch_helper {
   template <std::size_t... idx>
-  constexpr auto operator()(
-    const auto& v, 
-    const std::index_sequence<idx...>&) const 
+  constexpr auto operator()(const auto& v, const std::index_sequence<idx...>&) const 
   -> type_deduction_idx 
   {
     type_deduction_idx pipeline_seed = std::nullopt;
     return (
       pipeline_seed |
       ... |
-      [&]() { return type_switch_impl<idx, match_cases>{}(v); }
+      [&]() { return evaluate_switch_impl<idx, match_cases>{}(v); }
     );
   }
 };
 
 template <typename _switch>
-struct evaluate_type_switch;
+struct evaluate_switch;
 
 template <typename... cases>
-struct evaluate_type_switch<type_switch<cases...>> {
-  constexpr auto operator()(const auto& v) const -> 
-    type_deduction_res
+struct evaluate_switch<type_switch<cases...>> {
+  constexpr auto operator()(const auto& v) const -> type_deduction_res
   {
     auto res =
-      type_switch_helper<cases...>{}(
+      evaluate_switch_helper<cases...>{}(
         v, std::make_index_sequence<sizeof...(cases)>{}
       );
     if(!res)
